requirements.txt
protobuf==4.25.3 
grpcio==1.62.0 
grpcio-tools==1.62.0 
nltk>=3.8.1 